Fix canonical import pathing: add sys.path logic to src/ui/methods and output config UI
