+++ cfg/site-cfg/flame-cfg/flame-python/logik_projekt/cfg/logik_projekt_openclip_neat_video/config.xml
