+++ cfg/site-cfg/flame-cfg/flame-python/logik_projekt/src/core/logik_projekt_openclip.py
