+++ cfg/site-cfg/flame-cfg/flame-python/logik_projekt/src/ui/methods/pyside6_qt_flame_classes.py
@@ -2,170 +2,43 @@
 
 # -------------------------------------------------------------------------- #
 
-# File Name:        pyside6_qt_flame_modules.py
+# File Name:        pyside6_qt_flame_classes.py
 # Version:          1.0.3
 # Created:          2024-01-19
 # Modified:         2025-02-25
 
-# ========================================================================== #
-# This section imports the necessary modules and adds the modules directory
-# ========================================================================== #
-
-# Standard library imports
-import ast
-import datetime
-import functools
-import importlib.util
+# Ensure src package is discoverable for canonical imports
 import os
-import re
-import shutil
-import subprocess
 import sys
-import typing
-from typing import (
-    Union,
-    List,
-    Dict,
-    Optional,
-    Callable
-)
-import xml
-import xml.etree.ElementTree as ET
+current_file = os.path.abspath(__file__)
+src_parent = os.path.dirname(os.path.dirname(os.path.dirname(current_file)))
+if src_parent not in sys.path:
+	sys.path.insert(0, src_parent)
 
-# Third Party library imports
-try:
-    from PySide6 import (
-        QtWidgets,
-        QtCore,
-        QtGui,
-    )
-except ImportError:
-    from PySide2 import (
-        QtWidgets,
-        QtCore,
-        QtGui,
-    )
-
-# Get the directory path of the currently executing script
-current_script_dir = os.path.dirname(os.path.abspath(__file__))
-
-# Print the current_script_dir
-print(f"current_script_dir: {current_script_dir}")
-
-# The current_script_dir should be 'resources/flame/python/logik_projekt/openclip_tools/logik_projekt_openclip/scripts'
-# The modules directory should be 'resources/flame/python/logik_projekt/openclip_tools/logik_projekt_openclip/scripts/modules'
-# Check if this is true
-modules_dir = os.path.join(current_script_dir, 'modules')
-print(f"modules_dir: {modules_dir}")
-
-# Append the 'modules' directory to sys.path to access modules
-modules_dir = os.path.join(current_script_dir, 'modules')
-sys.path.append(modules_dir)
-
-# from pyside6_qt_flame_classes import (
-from modules.pyside6_qt_flame_classes import (
-# from modules.classes import (
-    pyside6_qt_button,
-    pyside6_qt_clickable_line_edit,
-    pyside6_qt_label,
-    pyside6_qt_line_edit,
-    pyside6_qt_list_widget,
-    pyside6_qt_message_window,
-    pyside6_qt_password_window,
-    pyside6_qt_preset_window,
-    pyside6_qt_progress_window,
-    pyside6_qt_push_button,
-    pyside6_qt_push_button_menu,
-    pyside6_qt_qdialog,
-    pyside6_qt_slider,
-    pyside6_qt_text_edit,
-    pyside6_qt_token_push_button,
-    pyside6_qt_tree_widget,
-    pyside6_qt_window,
-)
-
-# from pyside6_qt_flame_functions import (
-from modules.pyside6_qt_flame_functions import (
-# from modules.functions import (
-    pyside6_qt_get_shot_name,
-    pyside6_qt_print,
-    pyside6_qt_get_flame_version,
-    pyside6_qt_file_browser,
-    pyside6_qt_resolve_shot_name,
-    pyside6_qt_resolve_path_tokens,
-    pyside6_qt_refresh_hooks,
-    pyside6_qt_open_in_finder,
-    pyside6_qt_load_config,
-    pyside6_qt_save_config,
-)
-
-# from modules.functions.pyside6_qt_output_config_ui import (
-from modules.pyside6_qt_output_config_ui import (
-    pyside6_qt_output_config_ui as pyside6_qt_output_config_ui
-)
-
-# ============================== TEST ====================================== #
-
-# # Append the 'modules' directory to sys.path to access modules
-# if modules_dir not in sys.path:
-#     sys.path.append(modules_dir)
-
-# # Verify if the modules directory exists
-# if not os.path.exists(modules_dir):
-#     print(f"Error: modules_dir does not exist: {modules_dir}")
-# else:
-#     print(f"modules_dir exists: {modules_dir}")
-
-# # Import modules from the 'modules' directory
-# try:
-#     from modules.pyside6_qt_flame_classes import (
-#         pyside6_qt_button,
-#         pyside6_qt_clickable_line_edit,
-#         pyside6_qt_label,
-#         pyside6_qt_line_edit,
-#         pyside6_qt_list_widget,
-#         pyside6_qt_message_window,
-#         pyside6_qt_password_window,
-#         pyside6_qt_preset_window,
-#         pyside6_qt_progress_window,
-#         pyside6_qt_push_button,
-#         pyside6_qt_push_button_menu,
-#         pyside6_qt_qdialog,
-#         pyside6_qt_slider,
-#         pyside6_qt_text_edit,
-#         pyside6_qt_token_push_button,
-#         pyside6_qt_tree_widget,
-#         pyside6_qt_window,
-#     )
-
-#     from modules.pyside6_qt_flame_functions import (
-#         pyside6_qt_get_shot_name,
-#         pyside6_qt_print,
-#         pyside6_qt_get_flame_version,
-#         pyside6_qt_file_browser,
-#         pyside6_qt_resolve_shot_name,
-#         pyside6_qt_resolve_path_tokens,
-#         pyside6_qt_refresh_hooks,
-#         pyside6_qt_open_in_finder,
-#         pyside6_qt_load_config,
-#         pyside6_qt_save_config,
-#     )
-
-#     from modules.pyside6_qt_output_config_ui import (
-#         pyside6_qt_output_config_ui as pyside6_qt_output_config_ui
-#     )
-  
-# except ImportError as e:
-#     print(f"Error importing modules: {e}")
-
-# ============================== TEST ====================================== #
+from src.ui.widgets.classes.pyside6_qt_button import pyside6_qt_button
+from src.ui.widgets.classes.pyside6_qt_clickable_line_edit import pyside6_qt_clickable_line_edit
+from src.ui.widgets.classes.pyside6_qt_label import pyside6_qt_label
+from src.ui.widgets.classes.pyside6_qt_line_edit import pyside6_qt_line_edit
+from src.ui.widgets.classes.pyside6_qt_list_widget import pyside6_qt_list_widget
+from src.ui.widgets.classes.pyside6_qt_message_window import pyside6_qt_message_window
+from src.ui.widgets.classes.pyside6_qt_password_window import pyside6_qt_password_window
+from src.ui.widgets.classes.pyside6_qt_preset_window import pyside6_qt_preset_window
+from src.ui.widgets.classes.pyside6_qt_progress_window import pyside6_qt_progress_window
+from src.ui.widgets.classes.pyside6_qt_push_button import pyside6_qt_push_button
+from src.ui.widgets.classes.pyside6_qt_push_button_menu import pyside6_qt_push_button_menu
+from src.ui.widgets.classes.pyside6_qt_qdialog import pyside6_qt_qdialog
+from src.ui.widgets.classes.pyside6_qt_slider import pyside6_qt_slider
+from src.ui.widgets.classes.pyside6_qt_text_edit import pyside6_qt_text_edit
+from src.ui.widgets.classes.pyside6_qt_token_push_button import pyside6_qt_token_push_button
+from src.ui.widgets.classes.pyside6_qt_tree_widget import pyside6_qt_tree_widget
+from src.ui.widgets.classes.pyside6_qt_window import pyside6_qt_window
 
 # ========================================================================== #
 # This section defines how to handle the main script function.
 # ========================================================================== #
 
 # # If this script is executed as main:
-# # Call scripts for immediate execution
+# # Call functions for immediate execution
 # if __name__ == "__main__":
 
 # ========================================================================== #
@@ -284,7 +157,7 @@ from modules.pyside6_qt_output_config_ui import (
 # comments:              Replaced pyflame_print with pyside6_qt_print
 # -------------------------------------------------------------------------- #
 # version:               0.2.8
-# modified:              2024-05-17 - 13:49:59
+# modified:              2024-05-17 - 13:50:00
 # comments:              Replaced pyflame_refresh_hooks with pyside6_qt_refresh_hooks
 # -------------------------------------------------------------------------- #
 # version:               0.2.9
@@ -316,7 +189,7 @@ from modules.pyside6_qt_output_config_ui import (
 # comments:              Minor modification to Disclaimer.
 # -------------------------------------------------------------------------- #
 # version:               0.4.5
-# modified:              2024-06-04 - 17:38:53
+# modified:              2024-06-04 - 17:38:52
 # comments:              Added 'Smart Replace' option for render and write nodes
 # -------------------------------------------------------------------------- #
 # version:               0.4.6
