+++ cfg/site-cfg/flame-cfg/flame-python/logik_projekt/src/ui/widgets/classes/pyside6_qt_text_edit.py
