+++ cfg/site-cfg/flame-cfg/flame-python/logik_projekt/src/ui/widgets/functions/pyside6_qt_get_shot_name.py
