+++ cfg/site-cfg/flame-cfg/flame-python/logik_projekt/src/ui/widgets/functions/pyside6_qt_open_in_finder.py
