+++ pref/session-preferences/current_session-adsk.json
@@ -1,8 +1,8 @@
 {
-    "Name": "456_debug_python_scripts_2027_romeo",
-    "ProjectDir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo",
-    "SetupDir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo/setups",
-    "MediaDir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo/media",
+    "Name": "789_debug_flame_python_refactor_part_2_2027_romeo",
+    "ProjectDir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo",
+    "SetupDir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo/setups",
+    "MediaDir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo/media",
     "Resolution": {
         "FrameWidth": 1920,
         "FrameHeight": 1080,
