+++ pref/session-preferences/current_session-template.json
@@ -1,9 +1,9 @@
 {
-    "Template Serial Number: ": "456",
+    "Template Serial Number: ": "789",
     "Template Client Name: ": "debug",
-    "Template Campaign Name: ": "python scripts",
-    "Template Name: ": "456_debug_python_scripts",
-    "Template Description: ": "Debugging Logik Projekt Python Scripts for 2027",
+    "Template Campaign Name: ": "flame python refactor part 2",
+    "Template Name: ": "789_debug_flame_python_refactor_part_2",
+    "Template Description: ": "Complete refactoring of Logik Projekt Python Scripts",
     "Template Resolution: ": "1920 x 1080 x 1.0 | HD 1080 16:9",
     "Template Width: ": "1920",
     "Template Height: ": "1080",
