+++ pref/session-preferences/current_session-variables.json
@@ -8,14 +8,14 @@
     "flame_software_version": "2027.pr231",
     "flame_software_sanitized_name": "flame",
     "flame_software_sanitized_version": "2027",
-    "flame_projekt_name": "456_debug_python_scripts_2027_romeo",
-    "flame_projekt_nickname": "456_debug_python_scripts",
-    "flame_projekt_shotgun_name": "456_debug_python_scripts",
-    "flame_projekt_description": "Debugging Logik Projekt Python Scripts for 2027",
-    "flame_projekt_home": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo",
-    "flame_projekt_setups_dir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo/setups",
-    "flame_projekt_media_dir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo/media",
-    "flame_projekt_catalog_dir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo/catalog",
+    "flame_projekt_name": "789_debug_flame_python_refactor_part_2_2027_romeo",
+    "flame_projekt_nickname": "789_debug_flame_python_refactor_part_2",
+    "flame_projekt_shotgun_name": "789_debug_flame_python_refactor_part_2",
+    "flame_projekt_description": "Complete refactoring of Logik Projekt Python Scripts",
+    "flame_projekt_home": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo",
+    "flame_projekt_setups_dir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo/setups",
+    "flame_projekt_media_dir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo/media",
+    "flame_projekt_catalog_dir": "/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo/catalog",
     "flame_projekt_width": "1920",
     "flame_projekt_height": "1080",
     "flame_projekt_ratio": "1.7777778",
@@ -31,8 +31,8 @@
     "flame_projekt_cachef_id": 1185912,
     "flame_projekt_cachei": "Uncompressed (DPX)",
     "flame_projekt_cachei_id": 2,
-    "logik_projekt_name": "456_debug_python_scripts",
-    "logik_projekt_path": "/PROJEKTS/456_debug_python_scripts",
+    "logik_projekt_name": "789_debug_flame_python_refactor_part_2",
+    "logik_projekt_path": "/PROJEKTS/789_debug_flame_python_refactor_part_2",
     "logik_projekt_config_name": "logik-projekt",
     "logik_projekt_config_tree": "pref/site-prefs/default-prefs/logik-projekt-prefs/filesystem-tree.json",
     "logik_projekt_config_bookmarks": "pref/site-prefs/default-prefs/logik-projekt-prefs/cf_bookmarks.json",
