+++ pref/session-preferences/current_session-wiretap_template.xml
@@ -1,13 +1,13 @@
 <?xml version='1.0' encoding='utf-8'?>
 <Project>
-  <Name>456_debug_python_scripts_2027_romeo</Name>
-  <Nickname>456_debug_python_scripts</Nickname>
-  <ShotgunProjectName>456_debug_python_scripts</ShotgunProjectName>
-  <Description>Debugging Logik Projekt Python Scripts for 2027</Description>
-  <CatalogDir>/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo/catalog</CatalogDir>
-  <ProjectDir>/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo</ProjectDir>
-  <SetupDir>/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo/setups</SetupDir>
-  <MediaDir>/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/456_debug_python_scripts_2027_romeo/media</MediaDir>
+  <Name>789_debug_flame_python_refactor_part_2_2027_romeo</Name>
+  <Nickname>789_debug_flame_python_refactor_part_2</Nickname>
+  <ShotgunProjectName>789_debug_flame_python_refactor_part_2</ShotgunProjectName>
+  <Description>Complete refactoring of Logik Projekt Python Scripts</Description>
+  <CatalogDir>/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo/catalog</CatalogDir>
+  <ProjectDir>/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo</ProjectDir>
+  <SetupDir>/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo/setups</SetupDir>
+  <MediaDir>/Volumes/Samsung-T3-1TB/Autodesk/flame/projects/789_debug_flame_python_refactor_part_2_2027_romeo/media</MediaDir>
   <Version>2027.pr231</Version>
   <FrameWidth>1920</FrameWidth>
   <FrameHeight>1080</FrameHeight>
